refactor: consolidate API utilities
