+++ placeholder-main/lib/api.ts
@@ -1,6 +1,14 @@
 // lib/api.ts
 const BASE = (process.env.NEXT_PUBLIC_BACKEND_URL || 'http://127.0.0.1:8000').replace(/\/+$/, '');
 
+export type Agent = {
+  id: string;
+  handle: string;
+  display_name?: string;
+  avatar?: string;
+  resonance?: number;
+};
+
 export class ApiError extends Error {
   readonly status: number;
   readonly url: string;
@@ -40,6 +48,18 @@ async function request<T = any>(path: string, init: RequestInit = {}): Promise<T
   return res.json() as Promise<T>;
 }
 
+const AGENTS_BASE = process.env.NEXT_PUBLIC_API_BASE || '';
+
+async function json<T>(input: RequestInfo, init?: RequestInit): Promise<T> {
+  const res = await fetch(input, {
+    ...init,
+    headers: { 'Content-Type': 'application/json', ...(init?.headers || {}) },
+    cache: 'no-store',
+  });
+  if (!res.ok) throw new Error(`HTTP ${res.status}`);
+  return res.json() as Promise<T>;
+}
+
 /** GET /status */
 export function getStatus(signal?: AbortSignal) {
   return request('/status', { signal });
@@ -62,3 +82,22 @@ export function aiAssist(vibenodeId: number, prompt: string, token: string, sign
     signal,
   });
 }
+
+export const api = {
+  // Tries a few common agent endpoints; falls back gracefully
+  async listAgents(signal?: AbortSignal): Promise<Agent[]> {
+    if (!AGENTS_BASE) throw new Error('No API base set');
+    const paths = ['/agents', '/ai/personas', '/api/agents', '/api/personas'];
+    for (const p of paths) {
+      try {
+        const r = await json<any>(`${AGENTS_BASE}${p}`, { signal });
+        // normalize a little:
+        if (Array.isArray(r)) return r as Agent[];
+        if (Array.isArray(r?.results)) return r.results as Agent[];
+      } catch {
+        /* try next */
+      }
+    }
+    throw new Error('No agents endpoint found');
+  },
+};
